refactor(practice): extract pure metadata harvesters into engine layer
Move the ~304-line React-free block (harvestMetadata, harvestSummaryMetadata,
both per-mode harvester tables, all per-mode helpers, normalizeTarotIndex, and
the tarot date helpers) out of the ActiveRitualSession component and into a new
engine/harvestMetadata.ts module. The component re-imports the three public
symbols; the two harvester test files repoint at the new module. Also trims
three confirmed-dead fields (isSaving, wireMetadata, finishAndReset) from the
ActiveSession interface and return object. Pure mechanical move — behavior
byte-for-byte unchanged.
